requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
greenlet==3.3.2
idna==3.11
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dotenv==1.2.2
pywin32==311
requests==2.33.1
shiboken6==6.11.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
